Admin and User routers added

+++ routers/todos.py
@@ -5,6 +5,7 @@ from models import Todos
 from database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -16,6 +17,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[Session, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -24,30 +26,45 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.onwer_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     else:
         raise HTTPException(status_code=404, detail="Todo not found")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo_req: TodoRequest):
-    todo_model = Todos(**todo_req.dict())
+async def create_todo(user: user_dependency, db: db_dependency, todo_req: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = Todos(**todo_req.model_dump(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def update_todo(
+        user: user_dependency,
         db: db_dependency,
         todo_req: TodoRequest,
         todo_id: int = Path(gt=0)
     ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)\
+                  .filter(Todos.owner_id == user.get('id')).first())
     if todo_model is not None:
         raise HTTPException(status_code=404, detail="Todo not found")
 
@@ -60,10 +77,14 @@ async def update_todo(
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
 
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
